add opaque file
