+++ runtime/src/opaque.rs
@@ -1,0 +1,22 @@
+//! Opaque types. These are used by the CLI to instantiate machinery that don't need to know
+//! the specifics of the runtime. They can then be made to be agnostic over specific formats
+//! of data like extrinsics, allowing for them to continue syncing the network through upgrades
+//! to even the core data structures.
+
+use super::*;
+
+pub use sp_runtime::OpaqueExtrinsic as UncheckedExtrinsic;
+
+/// Opaque block header type.
+pub type Header = generic::Header<BlockNumber, BlakeTwo256>;
+/// Opaque block type.
+pub type Block = generic::Block<Header, UncheckedExtrinsic>;
+/// Opaque block identifier type.
+pub type BlockId = generic::BlockId<Block>;
+
+impl_opaque_keys! {
+	pub struct SessionKeys {
+		pub babe: Babe,
+		pub grandpa: Grandpa,
+	}
+}
